CMRNLP-23 fixing edits from readme, combining base_tool.py with loader.py, added comment into server.py added extra line to menifest files
