+++ tools/collections_embeddings/manifest.json
