+++ tools/geospatial_embeddings/manifest.json
